requirements.txt
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
paho-mqtt>=1.6.0
requests>=2.25.0
opencv-python>=4.5.0
Pillow>=8.0.0
transformers==4.53.3
accelerate>=0.20.0
timm
python-telegram-bot>=20.0
